moved userSetup.py to separate "install" folder so it doesn't run accidentally.
